Commit after merging a PR head into its base branch
This is a workaround to a LFS issue:

On protocubo/the-online-brt-planning-guide#24 we got `git lfs pull`
failures after a maintainer commit to the fork branch.

Aparently, `git lfs checkout` and/or `git lfs pull` have trouble dealing
with objects outside the current ref...  But we don't have any ideia so
far as to why did it work (in the same PR) before the maintainer commit.